+++ charge.go
@@ -32,7 +32,7 @@ type BankAccount struct {
 // ChargeRequest represents a Paystack charge request
 type ChargeRequest struct {
 	Email             string       `json:"email,omitempty"`
-	Amount            float32      `json:"amount,omitempty"`
+	Amount            int          `json:"amount,omitempty"`
 	Birthday          string       `json:"birthday,omitempty"`
 	Card              *Card        `json:"card,omitempty"`
 	Bank              *BankAccount `json:"bank,omitempty"`
